use a better vae to avoid NaN

+++ train_text_to_image_lora.py
@@ -604,7 +604,7 @@ def main(args):
     # The VAE is in float32 to avoid NaN losses.
     unet.to(accelerator.device, dtype=weight_dtype)
     if args.pretrained_vae_model_name_or_path is None:
-        vae.to(accelerator.device, dtype=weight_dtype)
+        vae.to(accelerator.device, dtype=torch.float32)
     else:
         vae.to(accelerator.device, dtype=weight_dtype)
     text_encoder_one.to(accelerator.device, dtype=weight_dtype)
